Moved more stuff out of CloudBlobCoreFileStorageService
