+++ src/NuGetGallery.Core/Services/GalleryCloudBlobFolderInformationProvider.cs
@@ -29,6 +29,70 @@ namespace NuGetGallery
             CoreConstants.Folders.PackagesContentFolderName,
         };
 
+        public string GetCacheControl(string folderName)
+        {
+            switch (folderName)
+            {
+                case CoreConstants.Folders.PackagesFolderName:
+                case CoreConstants.Folders.SymbolPackagesFolderName:
+                case CoreConstants.Folders.ValidationFolderName:
+                    return CoreConstants.DefaultCacheControl;
+
+                case CoreConstants.Folders.PackageBackupsFolderName:
+                case CoreConstants.Folders.UploadsFolderName:
+                case CoreConstants.Folders.SymbolPackageBackupsFolderName:
+                case CoreConstants.Folders.DownloadsFolderName:
+                case CoreConstants.Folders.PackageReadMesFolderName:
+                case CoreConstants.Folders.ContentFolderName:
+                case CoreConstants.Folders.RevalidationFolderName:
+                case CoreConstants.Folders.StatusFolderName:
+                case CoreConstants.Folders.UserCertificatesFolderName:
+                case CoreConstants.Folders.PackagesContentFolderName:
+                case CoreConstants.Folders.FlatContainerFolderName:
+                    return null;
+
+                default:
+                    throw new InvalidOperationException(
+                        string.Format(CultureInfo.CurrentCulture, "The folder name {0} is not supported.", folderName));
+            }
+        }
+
+        public string GetContentType(string folderName)
+        {
+            switch (folderName)
+            {
+                case CoreConstants.Folders.PackagesFolderName:
+                case CoreConstants.Folders.PackageBackupsFolderName:
+                case CoreConstants.Folders.UploadsFolderName:
+                case CoreConstants.Folders.ValidationFolderName:
+                case CoreConstants.Folders.SymbolPackagesFolderName:
+                case CoreConstants.Folders.SymbolPackageBackupsFolderName:
+                case CoreConstants.Folders.FlatContainerFolderName:
+                    return CoreConstants.PackageContentType;
+
+                case CoreConstants.Folders.DownloadsFolderName:
+                    return CoreConstants.OctetStreamContentType;
+
+                case CoreConstants.Folders.PackageReadMesFolderName:
+                    return CoreConstants.TextContentType;
+
+                case CoreConstants.Folders.ContentFolderName:
+                case CoreConstants.Folders.RevalidationFolderName:
+                case CoreConstants.Folders.StatusFolderName:
+                    return CoreConstants.JsonContentType;
+
+                case CoreConstants.Folders.UserCertificatesFolderName:
+                    return CoreConstants.CertificateContentType;
+
+                case CoreConstants.Folders.PackagesContentFolderName:
+                    return CoreConstants.OctetStreamContentType;
+
+                default:
+                    throw new InvalidOperationException(
+                        string.Format(CultureInfo.CurrentCulture, "The folder name {0} is not supported.", folderName));
+            }
+        }
+
         public bool IsPublicFolder(string folderName)
         {
             if (KnownPublicFolders.Contains(folderName))
