+++ src/NuGetGallery.Core/Services/ICloudBlobFolderInformationProvider.cs
@@ -9,6 +9,26 @@ namespace NuGetGallery
     /// </summary>
     public interface ICloudBlobFolderInformationProvider
     {
+        /// <summary>
+        /// Determines if specified folder is publicly accessible.
+        /// Used for creation of missing storage containers.
+        /// </summary>
+        /// <param name="folderName">Folder name to check.</param>
+        /// <returns>True if folder is publicly accessible, false otherwise.</returns>
         bool IsPublicFolder(string folderName);
+
+        /// <summary>
+        /// Determines the content type for files stored in the specified folder.
+        /// </summary>
+        /// <param name="folderName">Folder name.</param>
+        /// <returns>Content type string to be set for the blob.</returns>
+        string GetContentType(string folderName);
+
+        /// <summary>
+        /// Determines the cache control setting for the blobs stored in the specified folder.
+        /// </summary>
+        /// <param name="folderName">Folder name.</param>
+        /// <returns>Cache control string to be set for the blob being created.</returns>
+        string GetCacheControl(string folderName);
     }
 }
