+++ tests/NuGetGallery.Core.Facts/Services/GalleryCloudBlobFolderInformationProviderFacts.cs
@@ -1,0 +1,41 @@
+﻿// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using Xunit;
+
+namespace NuGetGallery
+{
+    public class GalleryCloudBlobFolderInformationProviderFacts
+    {
+        [Theory]
+        [FolderNamesData()]
+        public void ProvidesProperCacheControl(string folderName)
+        {
+            var target = new GalleryCloudBlobFolderInformationProvider();
+
+            var cacheControl = target.GetCacheControl(folderName);
+
+            if (folderName == CoreConstants.Folders.PackagesFolderName
+                || folderName == CoreConstants.Folders.SymbolPackagesFolderName
+                || folderName == CoreConstants.Folders.ValidationFolderName)
+            {
+                Assert.Equal(CoreConstants.DefaultCacheControl, cacheControl);
+            }
+            else
+            {
+                Assert.Null(cacheControl);
+            }
+        }
+
+        [Theory]
+        [FolderNamesData(includeContentTypes: true)]
+        public void ProvidesProperContentType(string folderName, string expectedContentType)
+        {
+            var target = new GalleryCloudBlobFolderInformationProvider();
+
+            var contentType = target.GetContentType(folderName);
+
+            Assert.Equal(expectedContentType, contentType);
+        }
+    }
+}
